+++ features/alerts/alertOnReindrake.js
@@ -2,19 +2,24 @@ import settings from "../../settings";
 import * as triggers from '../../constants/triggers';
 import * as seaCreatures from '../../constants/seaCreatures';
 import { NOTIFICATION_SOUND_SOURCE, OFF_SOUND_MODE } from "../../constants/sounds";
-import { isInSkyblock } from "../../utils/playerState";
+import { getWorldName, isInSkyblock } from "../../utils/playerState";
 import { getCatchTitle } from "../../utils/common";
+import { registerIf } from "../../utils/registers";
+import { JERRY_WORKSHOP } from "../../constants/areas";
 
 // Alert on any reindrake spawned in a lobby
 
-register("Chat", (event) => playAlertOnReindrake()).setCriteria(triggers.REINDRAKE_SPAWNED_BY_ANYONE_MESSAGE).setContains();
+registerIf(
+	register("Chat", (event) => playAlertOnReindrake()).setCriteria(triggers.REINDRAKE_SPAWNED_BY_ANYONE_MESSAGE).setContains(),
+	() => settings.alertOnAnyReindrakeCatch && isInSkyblock() && getWorldName() === JERRY_WORKSHOP
+);
 
 const reindrakeTrigger = triggers.RARE_CATCH_TRIGGERS.find(t => t.seaCreature == seaCreatures.REINDRAKE);
 const title = getCatchTitle(reindrakeTrigger.seaCreature, reindrakeTrigger.rarityColorCode);
 
 function playAlertOnReindrake() {
 	try {
-		if (!settings.alertOnAnyReindrakeCatch || !isInSkyblock()) {
+		if (!settings.alertOnAnyReindrakeCatch || !isInSkyblock() || getWorldName() !== JERRY_WORKSHOP) {
 			return;
 		}